workqueue: apply nrt logic for work which jumps to different wq
It ensures nrt logic for all conditions, otherwise when the work
jumps back, the nrt logic is broken, the workqueue core can't
find out the work pool that work is still running:

the work of wq1 is running on cpu A
the work jumps to wq2 and it is running on CPU B
the work jumps back to wq1 and it is running on CPU C

Now the work of wq1 is running in CPU A and C and the same time.

Jumping is not recomended, adds a warning.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>