workqueue: remove disabled wq_worker_waking_up()
When a worker is sleeping, its flags has WORKER_QUIT_CM, which means
worker->flags & WORKER_NOT_RUNNING is always non-zero, and which means
wq_worker_waking_up() is disabled.

so we removed wq_worker_waking_up(). and after this, the only access to
worker->flags which is not protected by "LI" is removed.

The patch also do these changes after removal:
1) because wq_worker_waking_up() is removed, we don't need schedule()
   before zapping nr_running in wq_unbind_fn(), and don't need to
   release/regain pool->lock.
2) the sanity check in worker_enter_idle() is changed to also check for
   unbound/disassociated pools. (because the above change and nr_running
   is expected always reliable in worker_enter_idle() now.)
3) xxxxxxxxxxxxxxxxxxxx

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>