workqueue: count the manager into pool->nr_idle
The existing manager is considered idle when too_many_workers()
or when the worker is trying to become to manager.

So it would be better to count the manager into pool->nr_idle.
And the pool->nr_idle becomes the manager-exclusion mechanism
instead of pool->manager_arb which will be removed in next patch.

All the usage-sites of the pool->nr_idle are carefully checked,
there are all accepted this change after the need_to_create_worker()
was re-implemented to use "list_empty(&pool->idle_list)".

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>