workqueue: remove the del_timer_sync()s in maybe_create_worker()
It is said in the document that the timer which is being
deleted by del_timer_sync() should not be restarted:
  Synchronization rules: Callers must prevent restarting of
  the timer, otherwise this function is meaningless.

Repeating timer may cause the del_timer_sync() spin longer,
or even spin forever in very very very very extreme condition.

It is not considered a bug for me, but it disobeys the document.

To fix it, we need:
  1) don't requeue the mayday timer when !need_to_create_worker()
       it is a preparation/cleaup step for the fix. Otherwise the timer
       will repeat forever if we only do the 2).
  2) remove the del_timer_sync()s in maybe_create_worker()
       Fix the problem. It is not enough if we only do the 1),
       need_to_create_worker() would be probably true when the time
       del_timer_sync() is called, the timer is still repeating.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>