workqueue: use dedicated creater kthread for all pools
There are some problems with the managers:
  1) The last idle worker prefer managing to processing.
     It is better that the processing of work items should be the first
     priority to make the whole system make progress earlier.
  2) managers among different pools can be parallel, but actually
     their major work are serialized on the kernel thread "kthreadd".
     These managers are sleeping and wasted when the system is lack
     of memory.

This patch introduces a dedicated creater kthread which offloads the
managing from the workers, thus every worker makes effort to process work
rather than create worker, and there is no manager wasting on sleeping
when the system is lack of memory.  This dedicated creater kthread causes
a little more work serialized than before, but it is acceptable.

Implement Detail:
  1) the creater_work:		do the creation, creation exclusion
  2) the cooldown timer:	cool down when fail
  3) the mayday timer:		dismiss itself when pool->nr_idle > 0
  4) the semantic for "pool->nr_idle == 0" active the creater_work, cooldown and mayday timer
  5) the routine of the creater_work:	create worker on two conditions
  6) worker_thread():		start management without unlock/wait/recheck/retry
  7) put_unbound_pool():	group destruction code by functionality
  8) init_workqueues():		init creater kthread earlier than pools/workers

  1) the creater_work
Every pool has a struct kthread_work creater_work to create worker, and
the dedicated creater kthread processes all these creater_works of
all pools. struct kthread_work has itself execution exclusion, so we don't
need the manager_arb to handle the creating exclusion any more.
put_unbound_pool() uses the flush_kthread_work() to synchronize with
the creating rather than uses the manager_arb.

  2) the cooldown timer
The cooldown timer is introduced to implement the cool-down mechanism
rather than to causes the creater to sleep.  When the create_worker()
fails, the cooldown timer is requested and it will restart the creater_work.

  3) the mayday timer
The mayday timer is changed so it doesn't restart itself when
pool->nr_idle > 0.  If it always restarts itself as before, we will add
a lot of complication in creater_work.  The creater_work will need to
call del_timer_sync() after successful creation and grab the pool->lock
to check and restart the timer when pool->nr_idle becomes zero.
We don't want that complication, let the timer dismiss itself.

  4) the semantic for "pool->nr_idle == 0"
Any moment when pool->nr_idle == 0, the pool must on the creating state.
The mayday timer must be active (pending or running) to ensure the mayday
can be sent, and at least one of the creater_work or the cooldown timer
must be active to ensure the creating is in progress or standby.  So the
last worker who causes the pool->nr_idle reduce to 0 has the responsibility
to kick the mayday timer and the creater_work. And may_start_working()
becomes misleading due to the meaning of "!pool->nr_idle" is changed and
any worker should start working in spite of the value of pool->nr_idle.
may_start_working() will be cleanup in the next patch with the intention
that the size of this patch can be shorter for reviewing.

  5) the routine of the creater_work
The creater_work creates a worker in these two conditions:
  A) pool->nr_idle == 0
     A new worker is needed to be created obviously even there is no
     work item pending.  The busy workers may sleep and the pool can't
     serves the future new work items if no new worker is standby or
     being created.
  B) pool->nr_idle == 1 && pool->nr_running == 0
     It should create a worker but not strictly needed since we still
     have a free idle worker and it can restart the creation when it goes
     to busy.  But if we don't create worker in this condition, this
     condition may occur frequently.  If a work item is queued and the
     last idle worker starts creater_work.  But if the work item is
     finished before the creater_work, this condition happens again,
     and it can happen again and again in this way.  So we had better
     to create a worker in this condition.
The creater_work will quit directly in other conditions.

  6) worker_thread()
There is no recheck nor retry when creating is required in worker_thread(),
it just kicks out the mayday timer and the creater work and goes to
process the work items directly.

  7) put_unbound_pool()
put_unbound_pool() groups code by functionality not by the name, it
stops creation activity (creater_work, cooldown_timer, mayday_timer)
at first and then stops idle workers and idle_tiemr.

  8) init_workqueues()
The struct kthread_worker kworker_creater is initialized earlier than
worker_pools in init_workqueues() so that kworker_creater_thread is
created than all early kworkers.  Although the early kworkers are not
depends on kworker_creater_thread, but this initialization order makes
the pid of kworker_creater_thread smaller than kworkers which
seems more smooth.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>